% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_balance_bagging_boosting.R
\name{InsExs}
\alias{InsExs}
\title{Generate synthetic instances from nearest neighbors}
\usage{
InsExs(instance, dataknns, numExs, nomAtt)
}
\arguments{
\item{instance}{A single instance to oversample from}

\item{dataknns}{Nearest neighbor instances}

\item{numExs}{Number of synthetic instances to generate}

\item{nomAtt}{Indices of nominal (factor) attributes}
}
\value{
A matrix containing generated synthetic instances
}
\description{
Generate synthetic instances from nearest neighbors
}
