% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_balance_bagging_boosting.R
\name{SmoteExs}
\alias{SmoteExs}
\title{Generate synthetic SMOTE instances for a minority class}
\usage{
SmoteExs(data, percOver, k)
}
\arguments{
\item{data}{A data frame containing the minority class instances}

\item{percOver}{Percentage of oversampling (e.g., 200 for 200\%)}

\item{k}{Number of nearest neighbors to consider}
}
\value{
A data frame containing synthetic instances for the minority class
}
\description{
Generate synthetic SMOTE instances for a minority class
}
